remove build-essential package after the ruby gems have been installed
The build-essentials Debian package is installed so that the ruby gems
can be compiled. This brings in gcc and numerous other packages that
are not needed in the final image. This patch removes the
build-essentials package, and its dependencies, after the ruby gems have
been compiled.

The output from `docker images` shows that the total image size goes
from 584.1 MB to 297.1 MB after this change.