Have VALIDATE_ONLY skip the install step
There is a use case in our CI environments where we need to quickly verify the configuration file. Because we must specify api/server/client we need to run a test against each type (e.g. /bin/start api). This infeasible when we have to run the /bin/install step as it would require us to run through the install process for api, client, and server.